+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Converter for Croatia (`hr`)
+- Converter for Germany, Mecklenburg-Western Pomerania
+- Converter for Germany, Saarland
+- Converter for Germany, Saxony
 
 ### Changed